+++ src/overlords/colonization/claimer.ts
@@ -1,6 +1,8 @@
 import {$} from '../../caching/GlobalCache';
+import {log} from '../../console/log';
 import {Roles, Setups} from '../../creepSetups/setups';
 import {Directive} from '../../directives/Directive';
+import {DirectiveColonizeShard} from 'directives/colony/colonize_shard';
 import {Pathing} from '../../movement/Pathing';
 import {OverlordPriority} from '../../priorities/priorities_overlords';
 import {profile} from '../../profiler/decorator';
@@ -9,11 +11,10 @@ import {Zerg} from '../../zerg/Zerg';
 import {Overlord} from '../Overlord';
 
 /**
- * Claim an unowned room
+ * Claim an unowned room or send claimer through portal
  */
 @profile
 export class ClaimingOverlord extends Overlord {
-
 	claimers: Zerg[];
 	directive: Directive;
 
@@ -25,39 +26,47 @@ export class ClaimingOverlord extends Overlord {
 
 	init() {
 		const amount = $.number(this, 'claimerAmount', () => {
-			if (this.room) { // if you have vision
-				if (this.room.my) { // already claimed
-					return 0;
-				} else { // don't ask for claimers if you can't reach controller
-					const pathablePos = this.room.creeps[0] ? this.room.creeps[0].pos
-															: Pathing.findPathablePosition(this.room.name);
-					if (!Pathing.isReachable(pathablePos, this.room.controller!.pos,
-											 _.filter(this.room.structures, s => !s.isWalkable))) {
-						return 0;
-					}
-				}
+			if (!this.room) return 1
+
+			// already claimed
+			if (this.room.my) return 0
+
+			// don't ask for claimers if you can't reach controller
+			const pathablePos = this.room.creeps[0] ? this.room.creeps[0].pos
+													: Pathing.findPathablePosition(this.room.name);
+			// if there is no controller, we want to colonize a new shard -> path to directive which is the portal
+			const pathDestination = this.room.controller ? this.room.controller.pos : this.pos
+			if (!Pathing.isReachable(pathablePos, pathDestination, _.filter(this.room.structures, s => !s.isWalkable))) {
+				log.warning(`Path for Directive ${this.directive.name} is not pathable`)
+				return 0
 			}
-			return 1; // otherwise ask for 1 claimer
+
+			return 1
 		});
 		const setup = this.colony.level > 4 ? Setups.infestors.fastClaim : Setups.infestors.claim;
 		this.wishlist(amount, setup);
 	}
 
 	private handleClaimer(claimer: Zerg): void {
-		if (claimer.room == this.room && !claimer.pos.isEdge) {
-			if (!this.room.controller!.signedByMe) {
-				// Takes care of an edge case where planned newbie zone signs prevents signing until room is reserved
-				if (!this.room.my && this.room.controller!.signedByScreeps) {
-					claimer.task = Tasks.claim(this.room.controller!);
-				} else {
-					claimer.task = Tasks.signController(this.room.controller!);
-				}
-			} else {
-				claimer.task = Tasks.claim(this.room.controller!);
-			}
-		} else {
+		if (claimer.room != this.room || claimer.pos.isEdge) {
 			claimer.goTo(this.pos, {pathOpts : {ensurePath: true, avoidSK: true}});
+			return
 		}
+
+		if (!this.room.controller && this.directive.directiveName == DirectiveColonizeShard.directiveName) {
+			// this is a portal room, just go on the portal
+			// the creep is already in the room
+			claimer.goTo(this.pos, {pathOpts: {maxRooms: 1}})
+			return
+		}
+
+		// Takes care of an edge case where planned newbie zone signs prevents signing until room is reserved
+		if (this.room.controller!.signedByMe || (!this.room.my && this.room.controller!.signedByScreeps)) {
+			claimer.task = Tasks.claim(this.room.controller!);
+			return
+		}
+
+		claimer.task = Tasks.signController(this.room.controller!);
 	}
 
 	run() {